+++ lib/v2/screens/dashboard/components/transaction_avatar.dart
