+++ lib/v2/screens/dashboard/components/transaction_details.dart
